Move the tree and iterators to common
