+++ piyavkin.anton/common/tree.hpp
@@ -21,35 +21,43 @@ namespace piyavkin
     Tree(InputIterator first, InputIterator second):
       Tree()
     {
-      while (first != second)
+      try
       {
-        insert(*first++);
+        while (first != second)
+        {
+          insert(*first++);
+        }
+      }
+      catch (...)
+      {
+        clear();
       }
     }
     Tree(std::initializer_list< val_type > il):
-      Tree()
-    {
-      auto it = il.begin();
-      while (it != il.end())
-      {
-        insert(*it++);
-      }
-    }
+      Tree(il.begin(), il.end())
+    {}
     Tree(const Tree& rhs):
       Tree()
     {
-      ConstTreeIterator< Key, T, Compare > it_right(rhs.root_);
-      ConstTreeIterator< Key, T, Compare > it_left(rhs.root_);
-      while (rhs.root_ && (it_right != rhs.cend() || it_left != --rhs.cbegin()))
+      try
       {
-        if (it_right != rhs.cend())
+        ConstTreeIterator< Key, T, Compare > it_right(rhs.root_);
+        ConstTreeIterator< Key, T, Compare > it_left(rhs.root_);
+        while (rhs.root_ && (it_right != rhs.cend() || it_left != --rhs.cbegin()))
         {
-          insert_impl(std::pair< Key, T >(*it_right++));
+          if (it_right != rhs.cend())
+          {
+            insert_impl(std::pair< Key, T >(*it_right++));
+          }
+          if (it_left != --rhs.cbegin())
+          {
+            insert_impl(std::pair< Key, T >(*it_left--));
+          }
         }
-        if (it_left != --rhs.cbegin())
-        {
-          insert_impl(std::pair< Key, T >(*it_left--));
-        }
+      }
+      catch (...)
+      {
+        clear();
       }
     }
     Tree(Tree&& rhs):
@@ -189,21 +197,21 @@ namespace piyavkin
     {
       return ConstTreeIterator< Key, T, Compare >(const_cast< detail::TreeNode< Key, T >* >(std::addressof(end_node_)));
     }
-    TreeReverseIterator< Key, T, Compare > rbegin() noexcept
+    std::reverse_iterator< TreeIterator< Key, T, Compare > > rbegin() noexcept
     {
-      return TreeReverseIterator< Key, T, Compare >(std::addressof(end_node_));
+      return std::reverse_iterator< TreeIterator< Key, T, Compare > >(end());
     }
-    TreeReverseIterator< Key, T, Compare > rend() noexcept
+    std::reverse_iterator< TreeIterator< Key, T, Compare > > rend() noexcept
     {
-      return TreeReverseIterator< Key, T, Compare >(before_min_.parent_);
+      return std::reverse_iterator< TreeIterator< Key, T, Compare > >(begin());
     }
-    ConstTreeReverseIterator< Key, T, Compare > crbegin() const noexcept
+    std::reverse_iterator< ConstTreeIterator< Key, T, Compare > > crbegin() const noexcept
     {
-      return ConstTreeReverseIterator< Key, T, Compare >(const_cast< detail::TreeNode< Key, T >* >(std::addressof(end_node_)));
+      return std::reverse_iterator< ConstTreeIterator< Key, T, Compare > >(cend());
     }
-    ConstTreeReverseIterator< Key, T, Compare > crend() const noexcept
+    std::reverse_iterator< ConstTreeIterator< Key, T, Compare > > crend() const noexcept
     {
-      return ConstTreeReverseIterator< Key, T, Compare >(before_min_.parent_);
+      return std::reverse_iterator< ConstTreeIterator< Key, T, Compare > >(cbegin());
     }
     TreeIterator< Key, T, Compare > find(const Key& key)
     {
@@ -221,7 +229,7 @@ namespace piyavkin
     TreeIterator< Key, T, Compare > upper_bound(const Key& key)
     {
       TreeIterator< Key, T, Compare > it = lower_bound_impl(key);
-      if (it.node_->val_type.first != key)
+      if (!cmp(it.node_->val_type.first, key) && !cmp(key, it.node_->val_type.first))
       {
         if (it != end())
         {
@@ -239,7 +247,7 @@ namespace piyavkin
     ConstTreeIterator< Key, T, Compare > upper_bound(const Key& key) const
     {
       ConstTreeIterator< Key, T, Compare > it = lower_bound(key);
-      if (it.node_->val_type.first != key)
+      if (equal_key(it.node_->val_type.first, key))
       {
         return it;
       }
@@ -416,7 +424,7 @@ namespace piyavkin
       }
       return count;
     }
-    std::pair< ConstTreeIterator< Key, T, Compare >, ConstTreeIterator< Key, T, Compare > > equil_range(const Key& key) const
+    std::pair< ConstTreeIterator< Key, T, Compare >, ConstTreeIterator< Key, T, Compare > > equal_range(const Key& key) const
     {
       ConstTreeIterator< Key, T, Compare > it = find_impl(key);
       if (it == cend())
@@ -426,23 +434,23 @@ namespace piyavkin
       ConstTreeIterator< Key, T, Compare > it2 = it;
       ConstTreeIterator< Key, T, Compare > temp = --it;
       ++it;
-      while (temp != --cbegin() && it.node_->val_type.first == temp.node_->val_type.first)
+      while (temp != --cbegin() && equal_key(it.node_->val_type.first, temp.node_->val_type.first))
       {
         --it;
         --temp;
       }
       temp = ++it2;
       --it2;
-      while (temp != cend() && it2.node_->val_type.first == temp.node_->val_type.first)
+      while (temp != cend() && equal_key(it2.node_->val_type.first, temp.node_->val_type.first))
       {
         ++it2;
         ++temp;
       }
       return std::pair< ConstTreeIterator< Key, T, Compare >, ConstTreeIterator< Key, T, Compare > >(it, ++it2);
     }
-    std::pair< TreeIterator< Key, T, Compare >, TreeIterator< Key, T, Compare > > equil_range(const Key& key)
+    std::pair< TreeIterator< Key, T, Compare >, TreeIterator< Key, T, Compare > > equal_range(const Key& key)
     {
-      auto pair = static_cast< const Tree< Key, T, Compare >& >(*this).equil_range(key);
+      auto pair = static_cast< const Tree< Key, T, Compare >& >(*this).equal_range(key);
       TreeIterator< Key, T, Compare > it1(const_cast< detail::TreeNode< Key, T >* >(pair.first.node_));
       TreeIterator< Key, T, Compare > it2(const_cast< detail::TreeNode< Key, T >* >(pair.second.node_));
       auto res = it1;
@@ -456,7 +464,7 @@ namespace piyavkin
     }
     size_t count(const Key& key) const
     {
-      std::pair< ConstTreeIterator< Key, T, Compare >, ConstTreeIterator< Key, T, Compare > > pair = equil_range(key);
+      std::pair< ConstTreeIterator< Key, T, Compare >, ConstTreeIterator< Key, T, Compare > > pair = equal_range(key);
       size_t sum = 0;
       while (pair.first != pair.second)
       {
@@ -577,8 +585,15 @@ namespace piyavkin
     }
     bool cmpNodeParent(const detail::TreeNode< Key, T >* node, const Key& key) const
     {
-      return (!root_ || node == root_ || (isLeftChild(node) && cmp_(key,node->parent_->val_type.first))
-        || (isRightChild(node) && cmp_(node->parent_->val_type.first, key)));
+      return (!root_ || node == root_ || isCorrectLeftChild(node, key) || isCorrectRightChild(node, key));
+    }
+    bool isCorrectLeftChild(const detail::TreeNode< Key, T >* node, const Key& key) const
+    {
+      return isLeftChild(node) && cmp_(key, node->parent_->val_type.first);
+    }
+    bool isCorrectRightChild(const detail::TreeNode< Key, T >* node, const Key& key) const
+    {
+      return isRightChild(node) && cmp_(node->parent_->val_type.first, key);
     }
     void zig(detail::TreeNode< Key, T >* node)
     {
@@ -757,11 +772,15 @@ namespace piyavkin
     ConstTreeIterator< Key, T, Compare > find_impl(const Key& key) const
     {
       ConstTreeIterator< Key, T, Compare > it = lower_bound_impl(key);
-      if (!it.node_ || it.node_->val_type.first != key)
+      if (!it.node_ || !equal_key(it.node_->val_type.first, key))
       {
         return cend();
       }
       return it;
+    }
+    bool equal_key(const Key& key1, const Key& key2) const
+    {
+      return !cmp_(key1, key2) && !cmp_(key2, key1);
     }
   };
 }
