+++ piyavkin.anton/common/treeiterator.hpp
@@ -9,15 +9,14 @@ namespace piyavkin
 {
   template< class Key, class T, class Compare >
   class Tree;
-  template< class Key, class T, class Compare >
-  class TreeReverseIterator;
   template< class Key, class T, class Compare = std::less< Key > >
   class TreeIterator: public std::iterator< std::bidirectional_iterator_tag, T >
   {
     using val_type = std::pair< Key, T >;
     friend class Tree< Key, T, Compare >;
-    friend class TreeReverseIterator< Key, T, Compare >;
   public:
+    using reference = std::pair< Key, T >&;
+    using pointer = std::pair< Key, T >*;
     TreeIterator():
       node_(nullptr)
     {}
@@ -80,11 +79,11 @@ namespace piyavkin
     {
       return !(node_ == rhs.node_);
     }
-    val_type* operator->()
+    pointer operator->()
     {
       return std::addressof(node_->val_type);
     }
-    val_type& operator*()
+    reference operator*()
     {
       return node_->val_type;
     }
@@ -106,15 +105,14 @@ namespace piyavkin
   class ConstLnrIterator;
   template< class Key, class T, class Compare >
   class ConstBreadthIterator;
-  template< class Key, class T, class Compare >
-  class ConstTreeReverseIterator;
   template< class Key, class T, class Compare = std::less< Key > >
   class ConstTreeIterator: public std::iterator< std::bidirectional_iterator_tag, T >
   {
     using val_type = std::pair< Key, T >;
     friend class Tree< Key, T, Compare >;
-    friend class ConstTreeReverseIterator< Key, T, Compare >;
   public:
+    using reference = const std::pair< Key, T >&;
+    using pointer = const std::pair< Key, T >*;
     ConstTreeIterator():
       node_(nullptr)
     {}
@@ -191,11 +189,11 @@ namespace piyavkin
     {
       return !(node_ == rhs.node_);
     }
-    const val_type* operator->() const
+    pointer operator->() const
     {
       return std::addressof(node_->val_type);
     }
-    const val_type& operator*() const
+    reference operator*() const
     {
       return node_->val_type;
     }
@@ -205,131 +203,8 @@ namespace piyavkin
       node_(node)
     {}
   };
-  template< class Key, class T, class Compare = std::less< Key > >
-  class TreeReverseIterator: public std::iterator< std::bidirectional_iterator_tag, T >
-  {
-    using val_type = std::pair< Key, T >;
-    friend class Tree< Key, T, Compare >;
-  public:
-    TreeReverseIterator() = default;
-    TreeReverseIterator(const TreeReverseIterator< Key, T, Compare >&) = default;
-    TreeReverseIterator< Key, T, Compare >& operator=(const TreeReverseIterator< Key, T, Compare >&) = default;
-    ~TreeReverseIterator() = default;
-    TreeReverseIterator< Key, T, Compare >& operator--()
-    {
-      ++it_;
-      return *this;
-    }
-    TreeReverseIterator< Key, T, Compare >& operator++()
-    {
-      --it_;
-      return *this;
-    }
-    TreeReverseIterator< Key, T, Compare > operator++(int)
-    {
-      TreeReverseIterator< Key, T, Compare > result(*this);
-      ++(*this);
-      return result;
-    }
-    TreeReverseIterator< Key, T, Compare > operator--(int)
-    {
-      TreeReverseIterator< Key, T, Compare > result(*this);
-      --(*this);
-      return result;
-    }
-    bool operator==(const TreeReverseIterator< Key, T, Compare >& rhs) const
-    {
-      return it_ == rhs.it_;
-    }
-    bool operator!=(const TreeReverseIterator< Key, T, Compare >& rhs) const
-    {
-      return !(it_ == rhs);
-    }
-    val_type* operator->()
-    {
-      return std::addressof(it_.node_->val_type);
-    }
-    val_type& operator*()
-    {
-      return it_.node_->val_type;
-    }
-    const val_type* operator->() const
-    {
-      return std::addressof(it_.node_->val_type);
-    }
-    const val_type& operator*() const
-    {
-      return it_.node_->val_type;
-    }
-  private:
-    TreeIterator< Key, T, Compare > it_;
-    explicit TreeReverseIterator(detail::TreeNode< Key, T >* node):
-      it_(node)
-    {}
-  };
   template< class Key, class T, class Compare >
   class ConstRnlIterator;
-  template< class Key, class T, class Compare = std::less< Key > >
-  class ConstTreeReverseIterator: public std::iterator< std::bidirectional_iterator_tag, T >
-  {
-    using val_type = std::pair< Key, T >;
-    friend class Tree< Key, T, Compare >;
-  public:
-    ConstTreeReverseIterator() = default;
-    ConstTreeReverseIterator(const ConstRnlIterator< Key, T, Compare >& it):
-      ConstTreeReverseIterator(it.stack.top())
-    {}
-    ConstTreeReverseIterator(const ConstTreeReverseIterator< Key, T, Compare >&) = default;
-    ConstTreeReverseIterator< Key, T, Compare >& operator=(const ConstTreeReverseIterator< Key, T, Compare >&) = default;
-    ConstTreeReverseIterator< Key, T, Compare >& operator=(const ConstRnlIterator< Key, T, Compare >& it)
-    {
-      it_ = it.stack.top();
-    }
-    ~ConstTreeReverseIterator() = default;
-    ConstTreeReverseIterator< Key, T, Compare >& operator--()
-    {
-      ++it_;
-      return *this;
-    }
-    ConstTreeReverseIterator< Key, T, Compare >& operator++()
-    {
-      --it_;
-      return *this;
-    }
-    ConstTreeReverseIterator< Key, T, Compare > operator++(int)
-    {
-      ConstTreeReverseIterator< Key, T, Compare > result(*this);
-      ++(*this);
-      return result;
-    }
-    ConstTreeReverseIterator< Key, T, Compare > operator--(int)
-    {
-      ConstTreeReverseIterator< Key, T, Compare > result(*this);
-      --(*this);
-      return result;
-    }
-    bool operator==(const ConstTreeReverseIterator< Key, T, Compare >& rhs) const
-    {
-      return it_ == rhs.it_;
-    }
-    bool operator!=(const ConstTreeReverseIterator< Key, T, Compare >& rhs) const
-    {
-      return !(it_ == rhs.it_);
-    }
-    const val_type* operator->() const
-    {
-      return std::addressof(it_.node_->val_type);
-    }
-    const val_type& operator*() const
-    {
-      return it_.node_->val_type;
-    }
-  private:
-    ConstTreeIterator< Key, T, Compare > it_;
-    explicit ConstTreeReverseIterator(detail::TreeNode< Key, T >* node):
-      it_(node)
-    {}
-  };
   template< class Key, class T, class Compare = std::less< Key > >
   class ConstLnrIterator: public std::iterator< std::bidirectional_iterator_tag, T >
   {
@@ -338,6 +213,8 @@ namespace piyavkin
     friend class ConstTreeIterator< Key, T, Compare >;
     friend class ConstRnlIterator< Key, T, Compare >;
   public:
+    using reference = const std::pair< Key, T >&;
+    using pointer = const std::pair< Key, T >*;
     ConstLnrIterator() = default;
     ConstLnrIterator(const ConstLnrIterator< Key, T, Compare >&) = default;
     ConstLnrIterator< Key, T, Compare >& operator=(const ConstLnrIterator< Key, T, Compare >&) = default;
@@ -410,11 +287,11 @@ namespace piyavkin
     {
       return !(*this == rhs);
     }
-    const val_type* operator->() const
+    pointer operator->() const
     {
       return std::addressof(stack.top()->val_type);
     }
-    const val_type& operator*() const
+    reference operator*() const
     {
       return stack.top()->val_type;
     }
